refactor(web): extract the lock policy and canvas-replacement coherence into hooks
Two self-contained coherence units leave SpatialEditor.tsx, verbatim:

- use-lock-policy.ts — whether a lock binds (host wired the seam), the
  isLocked/isEdgeLocked predicates, the selectable subset, and the two
  effects that retire selection/gesture state a lock arrival invalidates.
  Every downstream consumer (hit-testing, drag layers, snapping, keyboard
  verbs) takes the returned predicates, so lock policy has one definition.
- use-canvas-replacement.ts — the controlled-prop-swap useLayoutEffect
  (gesture abort/continue via the reducer's canvas-replaced event, plus
  retiring node/edge selections the new canvas no longer holds) together
  with the previous-value refs only it read. It stays a layout effect; the
  timing rationale moves with it. The call site sits after the setter
  declarations the inline effect could read lazily through closures.

No useState moves, so the editor-state-surface ledger is untouched.
Mutation-checked: forcing isLocked to false fails 8 node-lock browser
tests; making the replacement effect a no-op fails
selection-survives-external-delete; restores are green. Lock/marquee/
hit-target browser suites (29 tests) and the spatial-editor jsdom suite
(45 files / 491 tests) pass unchanged. SpatialEditor.tsx: 2717 -> 2622
lines (3364 at the start of this split).
